+++ R/board_registration.R
@@ -19,6 +19,7 @@ board_cache_path <- function() {
 #'
 #' @param name Optional name for this board, defaults to 'local'.
 #' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
 #'
 #' @seealso board_register
 #'
@@ -27,9 +28,11 @@ board_cache_path <- function() {
 #' board_register_local(cache = tempfile())
 #' @export
 board_register_local <- function(name = "local",
-                                 cache = board_cache_path()) {
+                                 cache = board_cache_path(),
+                                 ...) {
   board_register("local", name = name,
-                          cache = cache)
+                          cache = cache,
+                          ...)
 }
 
 #' Register GitHub Board
@@ -44,6 +47,7 @@ board_register_local <- function(name = "local",
 #' @param token Token to use when \code{GITHUB_PAT} is not specified.
 #' @param path The subdirectory in the repo where the pins will be stored.
 #' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
 #'
 #' @details
 #'
@@ -68,13 +72,15 @@ board_register_github <- function(name = "github",
                                   branch = "master",
                                   token = NULL,
                                   path = "",
-                                  cache = board_cache_path()) {
+                                  cache = board_cache_path(),
+                                  ...) {
   board_register("github", name = name,
                            repo = repo,
                            branch = branch,
                            token = token,
                            path = path,
-                           cache = cache)
+                           cache = cache,
+                           ...)
 }
 
 #' Register Kaggle Board
@@ -87,6 +93,7 @@ board_register_github <- function(name = "github",
 #'   be \code{NULL} if the \code{~/.kaggle/kaggle.json} file already exists.
 #' @param overwrite Should \code{~/.kaggle/kaggle.json} be overriden?
 #' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
 #'
 #' @seealso board_register
 #'
@@ -100,11 +107,13 @@ board_register_github <- function(name = "github",
 board_register_kaggle <- function(name = "kaggle",
                                   token = NULL,
                                   overwrite = FALSE,
-                                  cache = board_cache_path()) {
+                                  cache = board_cache_path(),
+                                  ...) {
   board_register("kaggle", name = name,
                            token = token,
                            overwrite = overwrite,
-                           cache = cache)
+                           cache = cache,
+                           ...)
 }
 
 #' Register RStudio Connect Board
@@ -118,6 +127,7 @@ board_register_kaggle <- function(name = "kaggle",
 #' @param key The RStudio Connect API key.
 #' @param output_files Should the output in an automated report create output files?
 #' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
 #'
 #' @seealso board_register
 #'
@@ -145,13 +155,15 @@ board_register_rsconnect <- function(name = "rsconnect",
                                      account = NULL,
                                      key = NULL,
                                      output_files = FALSE,
-                                     cache = board_cache_path()) {
+                                     cache = board_cache_path(),
+                                     ...) {
   board_register("rsconnect", name = name,
                               server = server,
                               account = account,
                               key = key,
                               output_files = output_files,
-                              cache = cache)
+                              cache = cache,
+                              ...)
 }
 
 #' Register Data TXT Board
@@ -161,7 +173,9 @@ board_register_rsconnect <- function(name = "rsconnect",
 #'
 #' @param name The name for this board, usually the domain name of the website.
 #' @param url Path to the \code{data.txt} file or path containing it.
+#' @param headers Optional list of headers to include or a function to generate them.
 #' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
 #'
 #' @seealso board_register
 #'
@@ -176,8 +190,138 @@ board_register_rsconnect <- function(name = "rsconnect",
 #' pin_find(board = "txtexample")
 #'
 #' @export
-board_register_datatxt <- function(name, url, cache = board_cache_path()) {
+board_register_datatxt <- function(name,
+                                   url,
+                                   headers = NULL,
+                                   cache = board_cache_path(),
+                                   ...) {
   board_register("datatxt", name = name,
                             url = url,
-                            cache = cache)
+                            headers = headers,
+                            cache = cache,
+                            ...)
+}
+
+#' Register S3 Board
+#'
+#' Wrapper with explicit parameters over \code{board_register()} to
+#' register an Amazon S3 bucket as a board.
+#'
+#' @param name Optional name for this board, defaults to 's3'.
+#' @param bucket The name of the Amazon S3 bucket. Defaults to the \code{AWS_BUCKET} environment
+#'   variable.
+#' @param key The key of the Amazon S3 bucket. Defaults to the \code{AWS_ACCESS_KEY_ID} environment
+#'   variable.
+#' @param secret The secret of the Amazon S3 bucket. Defaults to the \code{AWS_SECRET_ACCESS_KEY} environment
+#'   variable.
+#' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
+#'
+#' @details
+#'
+#' This function requires an Amazon S3 bucket to be manually created; otherwise,
+#' registering an S3 board will fail.
+#'
+#' @seealso board_register
+#'
+#' @examples
+#' \dontrun{
+#' # the following example requires an Amazon S3 API key
+#' board_register_s3(bucket = "s3bucket")
+#' }
+#' @export
+board_register_s3 <- function(name = "s3",
+                              bucket = Sys.getenv("AWS_BUCKET"),
+                              key = Sys.getenv("AWS_ACCESS_KEY_ID"),
+                              secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
+                              cache = board_cache_path(),
+                              ...) {
+  board_register("s3",
+                 name = name,
+                 bucket = bucket,
+                 key = key,
+                 secret = secret,
+                 cache = cache,
+                 ...)
+}
+
+#' Register Azure Board
+#'
+#' Wrapper with explicit parameters over \code{board_register()} to
+#' register a Microsoft Azure Storage Blob as a board.
+#'
+#' @param name Optional name for this board, defaults to 'azure'.
+#' @param container The name of the Azure Storage container. Defaults to the \code{AZURE_STORAGE_CONTAINER} environment
+#'   variable.
+#' @param account The account of the Azure Storage container. Defaults to the \code{AZURE_STORAGE_ACCOUNT} environment
+#'   variable.
+#' @param key The key of the Azure Storage container Defaults to the \code{AZURE_STORAGE_KEY} environment
+#'   variable.
+#' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
+#'
+#' @details
+#'
+#' This function requires an Azure Storage container to be manually created; otherwise,
+#' registering an Azire board will fail.
+#'
+#' @seealso board_register
+#'
+#' @examples
+#' \dontrun{
+#' # the following example requires an Azure Storage key
+#' board_register_azure(container = "pinscontainer",
+#'                      account = "pinsstorage",
+#'                      key = "abcabcabcabcabcabcabcabcabcab==")
+#' }
+#' @export
+board_register_azure <- function(name = "azure",
+                                 container = Sys.getenv("AZURE_STORAGE_CONTAINER"),
+                                 account = Sys.getenv("AZURE_STORAGE_ACCOUNT"),
+                                 key = Sys.getenv("AZURE_STORAGE_KEY"),
+                                 cache = board_cache_path(),
+                                 ...) {
+  board_register("azure",
+                 name = name,
+                 bucket = bucket,
+                 cache = cache,
+                 ...)
+}
+
+#' Register Google Cloud Board
+#'
+#' Wrapper with explicit parameters over \code{board_register()} to
+#' register a Google Cloud Storage container as a board.
+#'
+#' @param name Optional name for this board, defaults to 'gcloud'.
+#' @param bucket The name of the Google Cloud Storage bucket. Defaults to the \code{GCLOUD_STORAGE_BUCKET} environment
+#'   variable.
+#' @param token The access token of the Google Cloud Storage container. Defaults to use the Google Cloud SDK if configured.
+#' @param cache The local folder to use as a cache, defaults to \code{board_cache_path()}.
+#' @param ... Additional parameters required to initialize a particular board.
+#'
+#' @details
+#'
+#' This function requires a Google Cloud Storage container to be manually created; otherwise,
+#' registering a Google Cloud board will fail.
+#'
+#' @seealso board_register
+#'
+#' @examples
+#' \dontrun{
+#' # the following example requires the Google Cloud SDK to be configured
+#' board_register_gcloud(container = "gcloudcontainer")
+#' }
+#' @export
+board_register_gcloud <- function(name = "gcloud",
+                                  bucket = Sys.getenv("GCLOUD_STORAGE_BUCKET"),
+                                  token = NULL,
+                                  cache = board_cache_path(),
+                                  ...) {
+  board_register("gcloud",
+                 name = name,
+                 bucket = bucket,
+                 token = token,
+                 cache = cache,
+                 ...)
 }